test-requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymilvus[model]
milvus-lite
boto3
langchain-aws
openai
tiktoken
langchain
langchain-openai
langchain-community
langchain-core
pydantic
